Move planar_graphs -> planar
